[ 001 ] Initial commit
Initial commit of my OpenGL loader based on the xml files. For now it only load's and parses the xml, storing it results in a dictionary. It also has some basic settings, such as: OpenGL version, profile and option to search for gl.xml file. Now I'll focus on building the loader file; single-header library for loading OpenGL.