[ 010 ] Build scripts
I've written 2 Makefiles for this project: one is for the general generation of the loader, including an install rule for installing the loader on the system, and the second one for building sample code for, as of now, SDL3 and GLFW3. And at the topic of examples, I'm submitting another example, for SDL3.